+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change history for stripes-core
 
-## 3.6.0 (IN PROGRESS)
+## [3.6.0](https://github.com/folio-org/stripes-core/tree/v3.6.0) (2019-06-07)
+[Full Changelog](https://github.com/folio-org/stripes-core/compare/v3.5.1...v3.6.0)
 
 * Allow modules to be ingested by stripes-core as more than one type -- the `type` property of the `stripes` section is now replaced by `actsAs` which can be a module type string or an array of them. Also removes redundant copies of the icons, Okapi interfaces, and permissions in the `modules.<type>` structure exposed by the `ModulesContext` as they are kept elsewhere. (STCOR-148)
 * New settings icon.